refactor(federation,kernel): WAL stream/pipe become MetaStore-driven kernel primitives
Layering fix:

- Pipe / Stream are kernel primitives — they should know nothing about
  raft / federation, only the kernel-internal MetaStore HAL.
- Federation should DI ONE capability (the distributed MetaStore impl),
  not per-primitive backends (wal_stream_for_zone / wal_pipe_for_zone).

Concretely:

  - rust/kernel/src/abc/meta_store.rs: extend the trait with two
    raw-KV methods backing the stream-entries side table:
      * append_stream_entry(key, data)
      * get_stream_entry(key)
    Default impl returns NotSupported (LocalMetaStore / MemoryMetaStore
    have no replication) so callers fall back to in-memory pipe / stream
    backends; ZoneMetaStore (raft crate) overrides to propose
    Command::AppendStreamEntry / read sm.get_stream_entry.
  - rust/raft/src/zone_meta_store.rs: implement the two new methods.
  - rust/kernel/src/core/stream/wal.rs (new): WalStreamCore moves from
    the raft crate into the kernel crate (it is a kernel primitive
    composing whatever distributed MetaStore federation has DI'd).
    Holds Arc<dyn MetaStore>, no raft / WalConsensus reach.
  - rust/kernel/src/core/pipe/wal.rs (new): WalPipeCore composes
    WalStreamCore.  Per-replica head cursor + single-consumer FIFO
    semantics layered on top of the replicated entry log.  Same DI
    surface — Arc<dyn MetaStore>.
  - rust/raft/src/wal_stream_backend.rs: deleted.  Its WalConsensus
    trait and RaftWalConsensus impl are gone — the kernel-side
    MetaStore::{append_stream_entry, get_stream_entry} surface
    replaces them with no raft-internal types leaking out.
  - rust/kernel/src/hal/federation.rs: drop the wal_stream_for_zone /
    append_wal_entry / get_wal_entry trait methods.  Federation only
    DIs MetaStore via metastore_for_zone — WAL stream / pipe pull
    that store from the kernel's normal mount-route lookup.
  - rust/raft/src/federation_provider.rs: drop the matching impls.
  - rust/kernel/src/kernel/mod.rs::sys_setattr: both DT_PIPE and
    DT_STREAM io_profile=="wal" branches now resolve a MetaStore for
    the path's zone (federation_arc().metastore_for_zone) and
    construct a kernel-side WalStreamCore / WalPipeCore directly.  No
    per-primitive trait method needed.
  - rust/services/src/audit/mod.rs: AuditHook tests retarget mock
    storage from WalConsensus to MetaStore (single mock impl handles
    both the FileMetadata path and the new stream-entries surface).
  - rust/kernel/src/kernel/mod.rs::prepare_audit_stream: returns a
    concrete Arc<WalStreamCore> built from federation's MetaStore;
    AuditHook upcasts via `as Arc<dyn StreamBackend>`.

Net effect:
- Federation's responsibility shrinks to "install the distributed
  MetaStore for this zone" — no per-primitive DI.
- Pipe / Stream backends stay raft-unaware; replication is a
  metastore concern.
- WalConsensus trait disappears — the underlying KV is just MetaStore.
