+++ src/nexus/core/kernel_protocols.py
@@ -115,6 +115,8 @@ class MetaStore(Protocol):
     ) -> Any: ...
     def batch_get_content_ids(self, paths: Any) -> Any: ...
     def coherence_key(self) -> int | None: ...
+    def append_stream_entry(self, key: str, data: bytes) -> Any: ...
+    def get_stream_entry(self, key: str) -> Any: ...
 
 
 @runtime_checkable
@@ -153,9 +155,6 @@ class FederationProvider(Protocol):
     def list_zones(self, kernel: Any) -> list[str]: ...
     def metastore_for_zone(self, kernel: Any, zone_id: str) -> Any: ...
     def locks_for_zone(self, kernel: Any, zone_id: str) -> Any: ...
-    def wal_stream_for_zone(
-        self, kernel: Any, zone_id: str, stream_id: str, prefix: str
-    ) -> Any: ...
     def remote_read_blob(self, kernel: Any, zone_id: str, path: str, content_id: str) -> Any: ...
     def wire_mount(
         self, kernel: Any, parent_zone: str, mount_path: str, target_zone: str
@@ -173,12 +172,6 @@ class FederationProvider(Protocol):
     def lookup_share(self, kernel: Any, remote_path: str) -> Any: ...
     def zone_links_count(self, kernel: Any, zone_id: str) -> Any: ...
     def zone_cluster_info(self, kernel: Any, zone_id: str) -> Any: ...
-    def append_wal_entry(
-        self, kernel: Any, zone_id: str, entry_type: int, vfs_path: str, seq: int, entry: bytes
-    ) -> Any: ...
-    def get_wal_entry(
-        self, kernel: Any, zone_id: str, entry_type: int, vfs_path: str, seq: int
-    ) -> Any: ...
 
 
 @runtime_checkable
